backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
cryptography==41.0.7
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
flasgger==0.9.7.1
pdfminer.six==20221105
PyMuPDF==1.23.26
spacy==3.7.4
python-docx==1.1.0
